fixed up the data handler protocol to have the right delcartion of methods that are actually used also cleaned up all the integer values to intValue.

+++ SurfShackMobile/SubSettingColorViewController.m
@@ -80,7 +80,7 @@
 {
     [tableView deselectRowAtIndexPath:indexPath animated:YES];
     
-    int selectedRow = indexPath.row;
+    int selectedRow = (int)indexPath.row;
     
     [PreferenceFactory setColorPreference:[tableData objectAtIndex:indexPath.row]];
